Keyboard
Replace the collection view footer.  The footer was the keyboard, but only the top row would register key taps.  Several methods were attempted to get the keyboard to work properly.  I replaced the footer with a UIView and that fixes the issue.

Performed some cleanup around the layout and actions.

Code cleanup

Changes to be committed:
	modified:   Wordle/Board/BoardCell.swift
	modified:   Wordle/Board/BoardViewController.swift
	modified:   Wordle/Board/LetterView.swift
	modified:   Wordle/Keyboard/KeyboardCell.swift
	modified:   Wordle/Keyboard/KeyboardView.swift